Move HostUtils functionality to HostObject
